Separating chords classes in different files
