+++ Client/PROJECTS/Jamtaba-common.pri
@@ -77,6 +77,10 @@ HEADERS += gui/chords/ChordLabel.h
 HEADERS += gui/chords/ChordsWidget.h
 HEADERS += gui/chords/ChordsPanel.h
 HEADERS += gui/chords/ChordProgression.h
+HEADERS += gui/chords/ChordProgressionMeasure.h
+HEADERS += gui/chords/ChordsProgressionParser.h
+HEADERS += gui/chords/ChatChordsProgressionParser.h
+HEADERS += gui/chords/Chord.h
 HEADERS += gui/LocalTrackView.h
 HEADERS += gui/JamRoomViewPanel.h
 HEADERS += ninjam/UserChannel.h
@@ -132,11 +136,14 @@ SOURCES += gui/UserNameDialog.cpp
 SOURCES += gui/MainWindow.cpp
 SOURCES += log/logging.cpp
 SOURCES += gui/widgets/CustomTabWidget.cpp
-SOURCES += gui/chords/ChordLabel.cpp
 SOURCES += gui/BpiUtils.cpp
 SOURCES += gui/chords/ChordsWidget.cpp
 SOURCES += gui/chords/ChordsPanel.cpp
 SOURCES += gui/chords/ChordProgression.cpp
+SOURCES += gui/chords/ChordProgressionMeasure.cpp
+SOURCES += gui/chords/Chord.cpp
+SOURCES += gui/chords/ChordLabel.cpp
+SOURCES += gui/chords/ChatChordsProgressionParser.cpp
 SOURCES += Configurator.cpp
 SOURCES += persistence/UsersDataCache.cpp
 SOURCES += gui/widgets/IntervalChunksDisplay.cpp
